⚡️ Refactor : 컴포넌트 분리 및 메타태그 설정
