+++ src/lib/api/ElectionApi.ts
@@ -2,7 +2,7 @@ import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { axiosInstance } from "./axiosInstance";
 
 export interface CreateDepartLocationRequest {
-  memberId: string;
+  //memberId: string;
   departLocationName: string;
   latitude: number;
   longitude: number;
@@ -42,6 +42,7 @@ export const createDepartLocation = async (
       },
     }
   );
+  console.log("출발 장소 response:", res.data);
   return res.data;
 };
 
@@ -54,6 +55,7 @@ export const voteMiddleLocation = async (
   scheduleMemberId: number,
   body: { locationId: number; scheduleId: number }
 ) => {
+  console.log("투표 body:", body);
   const res = await axiosInstance.post(
     `/schedules/suggested-locations/vote/${scheduleMemberId}`,
     body
@@ -128,11 +130,11 @@ export const useVoteDepartLocation = () => {
     },
   });
 };
-
+/*
 //세부 일정 api 연결
 export const getSchedule = async (scheduleId: string) => {
   const res = await axiosInstance.get(`/schedules/show/${scheduleId}`);
-  return res.data.data;
+  return res.data;
 };
 
 export const useSchedule = (scheduleId: string) => {
@@ -144,6 +146,7 @@ export const useSchedule = (scheduleId: string) => {
     refetchOnWindowFocus: false,
   });
 };
+*/
 
 //즐겨찾기 장소 조회
 export const getFavoriteLocation = async () => {
